fixing up the new revised parsing logic

+++ src/generateCSS/utils/generate_utils/createFileAndAppendCSSRules.js
@@ -5,5 +5,9 @@ export function createFileAndAppendCSSRules(propertyName, fileContent) {
   const outputDir = path.join(global.__basedir, "dist", "css");
   const fileName = propertyName + ".css";
   const filePath = path.join(outputDir, fileName);
-  fs.writeFileSync(filePath, fileContent, "utf8");
+  if (fs.existsSync(filePath)) {
+    fs.appendFileSync(filePath, fileContent, "utf8");
+  } else {
+    fs.writeFileSync(filePath, fileContent, "utf8");
+  }
 }
